chalicelib/requirements.txt
a2wsgi==1.6.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.89.1
frozenlist==1.3.3
idna==3.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.26.4
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
